pyproject.toml
[project]
name = "rds-agent"
version = "0.2.0"
description = "RDS智能问答助手和智能运维助手 - MySQL数据库智能Agent (Django + Celery)"
authors = [
    { name = "RDS Agent Team" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # LangChain核心
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",

    # 本地模型支持
    "ollama>=0.1.0",

    # 数据库连接
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "dbutils>=3.0.0",
    "cryptography>=42.0.0",

    # Django框架
    "django>=4.2.0",
    "djangorestframework>=3.15.0",

    # Celery
    "celery>=5.3.0",
    "redis>=5.0.0",
    "django-celery-beat>=2.6.0",

    # MySQL数据库驱动
    "mysqlclient>=2.2.0",

    # 数据模型和配置 (保留用于诊断模块)
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # 向量知识库
    "chromadb>=0.5.0",

    # CLI交互
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",

    # HTTP客户端
    "httpx>=0.27.0",

    # 工具
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-celery>=1.0.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "black>=24.0.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
rds-agent = "rds_agent.main:main"
rds-agent-cli = "rds_agent.cli:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rds_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "rds_agent.django_project.settings"
django_debug_mode = true
addopts = "-v --cov=src/rds_agent --cov-report=term-missing --cov-report=html"
filterwarnings = [
    "ignore::DeprecationWarning",
]